+++ qoqo-qryd/src/tweezer_devices.rs
@@ -154,17 +154,22 @@ impl TweezerDeviceWrapper {
             .as_str()
     }
 
-    /// Switch to a different pre-defined layout.
+    /// Switch to a different pre-defined Layout.
+    ///
+    /// It is updated only if the given Layout name is present in the device's
+    /// Layout register. If the qubit -> tweezer mapping is empty, it is
+    /// trivially populated by default.
     ///
     /// Args:
-    ///     layout_number (str): The number index of the new layout
+    ///     layout_number (str): The number index of the new Layout.
+    ///     with_trivial_map (bool): Whether the qubit -> tweezer mapping should be trivially populated. Defaults to true.
     ///
     /// Raises:
     ///     PyValueError
-    #[pyo3(text_signature = "(name, /)")]
-    pub fn switch_layout(&mut self, name: &str) -> PyResult<()> {
+    #[pyo3(text_signature = "(name, with_trivial_map, /)")]
+    pub fn switch_layout(&mut self, name: &str, with_trivial_map: Option<bool>) -> PyResult<()> {
         self.internal
-            .switch_layout(name)
+            .switch_layout(name, with_trivial_map)
             .map_err(|err| PyValueError::new_err(format!("{:}", err)))
     }
 
@@ -200,6 +205,20 @@ impl TweezerDeviceWrapper {
                 Ok(mapping) => Ok(mapping.into_py_dict(py).into()),
                 Err(err) => Err(PyValueError::new_err(format!("{:}", err))),
             }
+        })
+    }
+
+    /// Get the qubit -> tweezer mapping of the device.
+    ///
+    /// Returns:
+    ///     dict[int, int]: The qubit -> tweezer mapping.
+    ///     None: The mapping is empty.
+    pub fn get_qubit_to_tweezer_mapping(&self) -> Option<PyObject> {
+        Python::with_gil(|py| -> Option<PyObject> {
+            self.internal
+                .qubit_to_tweezer
+                .as_ref()
+                .map(|mapping| mapping.into_py_dict(py).into())
         })
     }
 
@@ -481,7 +500,7 @@ impl TweezerDeviceWrapper {
             .map_err(|_| PyValueError::new_err("Input cannot be deserialized to TweezerDevice"))?;
         if let Some(layout) = &internal.default_layout {
             let _ = internal
-                .switch_layout(&layout.to_string())
+                .switch_layout(&layout.to_string(), None)
                 .map_err(|err| PyValueError::new_err(format!("{:}", err)));
         }
         Ok(TweezerDeviceWrapper { internal })
@@ -497,11 +516,15 @@ impl TweezerDeviceWrapper {
 
     /// Returns the number of total tweezer positions in the device.
     ///
+    /// Arguments:
+    ///     layout_name (Optional[str]): The name of the layout to reference. Defaults to the current layout.
+    ///
     /// Returns:
     ///     int: The number of tweezer positions in the device.
-    pub fn number_tweezer_positions(&self) -> PyResult<usize> {
+    #[pyo3(text_signature = "(layout_name, /)")]
+    pub fn number_tweezer_positions(&self, layout_name: Option<String>) -> PyResult<usize> {
         self.internal
-            .number_tweezer_positions()
+            .number_tweezer_positions(layout_name)
             .map_err(|err| PyValueError::new_err(format!("{:}", err)))
     }
 
@@ -657,17 +680,22 @@ impl TweezerMutableDeviceWrapper {
             .map_err(|err| PyValueError::new_err(format!("{:}", err)))
     }
 
-    /// Switch to a different pre-defined layout.
+    /// Switch to a different pre-defined Layout.
+    ///
+    /// It is updated only if the given Layout name is present in the device's
+    /// Layout register. If the qubit -> tweezer mapping is empty, it is
+    /// trivially populated by default.
     ///
     /// Args:
-    ///     layout_number (str): The number index of the new layout
+    ///     layout_number (str): The number index of the new Layout.
+    ///     with_trivial_map (bool): Whether the qubit -> tweezer mapping should be trivially populated. Defaults to true.
     ///
     /// Raises:
     ///     PyValueError
-    #[pyo3(text_signature = "(name, /)")]
-    pub fn switch_layout(&mut self, name: &str) -> PyResult<()> {
+    #[pyo3(text_signature = "(name, with_trivial_map, /)")]
+    pub fn switch_layout(&mut self, name: &str, with_trivial_map: Option<bool>) -> PyResult<()> {
         self.internal
-            .switch_layout(name)
+            .switch_layout(name, with_trivial_map)
             .map_err(|err| PyValueError::new_err(format!("{:}", err)))
     }
 
@@ -703,6 +731,20 @@ impl TweezerMutableDeviceWrapper {
                 Ok(mapping) => Ok(mapping.into_py_dict(py).into()),
                 Err(err) => Err(PyValueError::new_err(format!("{:}", err))),
             }
+        })
+    }
+
+    /// Get the qubit -> tweezer mapping of the device.
+    ///
+    /// Returns:
+    ///     dict[int, int]: The qubit -> tweezer mapping.
+    ///     None: The mapping is empty.
+    pub fn get_qubit_to_tweezer_mapping(&self) -> Option<PyObject> {
+        Python::with_gil(|py| -> Option<PyObject> {
+            self.internal
+                .qubit_to_tweezer
+                .as_ref()
+                .map(|mapping| mapping.into_py_dict(py).into())
         })
     }
 
@@ -1000,11 +1042,15 @@ impl TweezerMutableDeviceWrapper {
 
     /// Returns the number of total tweezer positions in the device.
     ///
+    /// Arguments:
+    ///     layout_name (Optional[str]): The name of the layout to reference. Defaults to the current layout.
+    ///
     /// Returns:
     ///     int: The number of tweezer positions in the device.
-    pub fn number_tweezer_positions(&self) -> PyResult<usize> {
+    #[pyo3(text_signature = "(layout_name, /)")]
+    pub fn number_tweezer_positions(&self, layout_name: Option<String>) -> PyResult<usize> {
         self.internal
-            .number_tweezer_positions()
+            .number_tweezer_positions(layout_name)
             .map_err(|err| PyValueError::new_err(format!("{:}", err)))
     }